+++ packages/rust-ipns/CHANGELOG.md
@@ -1,2 +1,7 @@
+# 0.2.0 [unreleased]
+- chore: Update libp2p to 0.53 [PR 116]
+
+[PR 116]:  https://github.com/dariusc93/rust-ipfs/pull/116
+
 # 0.1.1
 - Initial Release